+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v0.7.0 (TBD)
  - [Import] Allow user to keep dicom files in the selected import directory
  - [Modeling] Allow user to open ML model without explicitly generating an MVR model
  - [Query] Fix for [Issue #37](https://github.com/cutright/DVH-Analytics/issues/37)
+ - [Query] Proper min and max label updates with units in numerical query dialog
  - [RadBio] Fix for [Issue #34](https://github.com/cutright/DVH-Analytics/issues/34)
  - [Misc] Corrected units for beam perimeter and area